Cleaned up Components.js
