+++ src/js/shot-generator-new/utils/preventDefault.js
@@ -1,0 +1,6 @@
+
+const preventDefault = (fn, ...args) => e => {
+    e.preventDefault()
+    fn(e, ...args)
+  }
+export { preventDefault }
